Refactor: Split domain module into focused files
- Created events.py for Event class
- Created interfaces.py for IActionData, ILibrary, IContainer, etc.
- Created ui_interfaces.py for IBindingModel
- Updated __init__.py to re-export for backward compatibility
- Better separation of concerns (SRP - SOLID)
- Improved maintainability and discoverability
